test(rust_streams): add unit tests for Linux splice optimization and fallback
- Added new test file `test_rust_splice.py` with tests verifying Linux splice() zero-copy path and fallback to read/write.
- Removed deprecated threaded pump helper from `test_rust_streams.py`.
- Updated test coverage to handle large transfers, fallback scenarios, and broken pipe draining.
- Improved test reliability and clarity for Linux-specific behavior.
